payments/pyproject.toml
[project]
name = "payments"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.117.1,<0.118.0)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<1.0.0)",
    "greenlet (>=3.0.0,<4.0.0)",
    "pulsar-client[avro] (>=3.8.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"